[IMP] mail: queue message before posting them
Before this commit, composer would freeze while a message is sending. This
commit introduce a message queue that will store the pending message.

Task-2390637
